# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación Atlas",
        "description": "Documentación WPARK - API",
        "contact": {
            "email": "lchirinos@acercapartners.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Users List",
                "description": "List of users",
                "operationId": "indexUsers",
                "responses": {
                    "200": {
                        "description": "User list Successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create New User",
                "description": "Create New User",
                "operationId": "storeUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create New User"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show User Details",
                "description": "Show User Details",
                "operationId": "showUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show User Details"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update User",
                "description": "Update User",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update User successfully"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "422": {
                        "$ref": "#/components/responses/UnprocessableEntity"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete User",
                "description": "Delete User",
                "operationId": "destroyUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete User successfully"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Users"
                ],
                "summary": "Restore User",
                "description": "Restore User",
                "operationId": "restoreUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User restored successfully"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/me": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Show Login User Details",
                "description": "Show Login User Details",
                "operationId": "meUser",
                "responses": {
                    "200": {
                        "description": "Show Login User Details"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "403": {
                        "$ref": "#/components/responses/Forbidden"
                    },
                    "500": {
                        "$ref": "#/components/responses/InternalServerError"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserStoreRequest": {
                "title": "User Store Request",
                "description": "User Store request body data",
                "required": [
                    "name",
                    "email",
                    "username",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string",
                        "maxLength": 75,
                        "example": "Javier"
                    },
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "maxLength": 255,
                        "example": "jgbacerca@gmail.com"
                    },
                    "username": {
                        "description": "Nombre de usuario",
                        "type": "string",
                        "maxLength": 255,
                        "example": "javier.garrido"
                    },
                    "password": {
                        "description": "Contraseña del usuario",
                        "type": "string",
                        "maxLength": 100,
                        "example": "jgbTest6;"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserStoreRequest"
                }
            },
            "UserUpdateRequest": {
                "title": "User Update Request",
                "description": "User Update request body data",
                "required": [
                    "name",
                    "email",
                    "username"
                ],
                "properties": {
                    "name": {
                        "description": "Nombre del usuario",
                        "type": "string",
                        "maxLength": 75,
                        "example": "Javier"
                    },
                    "surname": {
                        "description": "Apellidos del usuario",
                        "type": "string",
                        "maxLength": 255,
                        "example": "Garrido Barroso"
                    },
                    "email": {
                        "description": "Email del usuario",
                        "type": "string",
                        "maxLength": 255,
                        "example": "jgbacerca@gmail.com"
                    },
                    "username": {
                        "description": "Nombre de usuario",
                        "type": "string",
                        "maxLength": 255,
                        "example": "javier.garrido"
                    },
                    "admin_pin": {
                        "description": "Pin del administrador",
                        "type": "integer",
                        "maxLength": 10,
                        "example": "2584639875"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserUpdateRequest"
                }
            },
...
        },
        "responses": {
...
            "Forbidden": {
                "description": "Permission Error."
            },
            "InternalServerError": {
                "description": "Internal Server Error"
            },
            "NotFound": {
                "description": "Resource Not Found."
            },
            "Unauthorized": {
                "description": "Unauthorized Error."
            },
            "UnprocessableEntity": {
                "description": "Validation Error."
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "id, eg; 1",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        []
    ]
}